requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
requests==2.32.5
python-dotenv==1.2.1
mysql-connector-python==9.5.0
pydantic==2.12.4
starlette==0.49.3
